requirements.txt
pydantic>=2.5.0
fastapi>=0.110.0
uvicorn>=0.24.0
aiosqlite>=0.19.0
jinja2>=3.1.2
python-multipart>=0.0.6
pypdf>=3.17.0
striprtf>=0.0.26
extract-msg>=0.41.1
beautifulsoup4>=4.12.3
python-dateutil>=2.8.2
httpx>=0.26.0
pytest>=8.1.0
email-validator>=2.2.0
